update name file and UI
